Separate inference out of the main class
Also, rename about every variable...
