+++ autogp/__init__.py
@@ -1,4 +1,5 @@
-from .gaussian_process import GaussianProcess
+from .gp import GaussianProcess
 from . import datasets
-from . import likelihoods
-from . import kernels
+from . import lik
+from . import cov
+from . import inf
