+++ autogp/cov/__init__.py
@@ -1,0 +1,3 @@
+from .cov import Cov
+from .cov_se import SquaredExponential
+from .arc_cosine import ArcCosine
