+++ autogp/inf/__init__.py
@@ -1,0 +1,2 @@
+from .inf import Inference
+from .inf_vi import VariationalInference
