--- autogp/kernels/__init__.py
@@ -1,3 +1,0 @@
-from .kernel import Kernel
-from .radial_basis import RadialBasis
-from .arc_cosine import ArcCosine
