+++ autogp/lik/logistic.py
