+++ autogp/lik/softmax.py
