+++ autogp/util/__init__.py
@@ -1,7 +1,7 @@
 from .normals import CholNormal
 from .normals import DiagNormal
-from .ops import tri_vec_shape
-from .ops import vec_to_tri
+from .util import tri_vec_shape
+from .util import vec_to_tri
 from .util import ceil_divide
 from .util import get_flags
 from .util import log_cholesky_det
