+++ autogp/util/util.py
@@ -1,4 +1,5 @@
 import tensorflow as tf
+import numpy as np
 
 
 def _merge_and_separate(a, b, func):
@@ -140,6 +141,30 @@ def broadcast(tensor, tensor_with_target_shape):
         tile_multiples = tf.concat([target_shape[0:-input_rank], [1] * input_rank], axis=0)
     input_with_expanded_dims = tf.reshape(tensor, expand_dims_shape)
     return tf.tile(input_with_expanded_dims, tile_multiples)
+
+
+def tri_vec_shape(N):
+    """
+    Compute the length of a vector that contains all elements of a triangle with dimensions N.
+    """
+    return [N * (N + 1) // 2]
+
+
+def vec_to_tri(vectors):
+    """
+    Takes a DxM tensor vectors and maps it to a D x matrix_size x matrix_size
+    tensor where the lower triangle of each matrix_size x matrix_size matrix
+    is constructed by unpacking each M-vector.
+    """
+    M = vectors.shape[1].value
+    N = int(np.floor(0.5 * np.sqrt(M * 8. + 1) - 0.5))
+    assert N * (N + 1) == 2 * M  # check M is a valid triangle number.
+    indices = tf.constant(np.stack(np.tril_indices(N)).T, dtype=tf.int32)
+
+    def vec_to_tri_vector(vector):
+        return tf.scatter_nd(indices=indices, shape=[N, N], updates=vector)
+
+    return tf.map_fn(vec_to_tri_vector, vectors)
 
 
 def get_flags():
