+++ experiments/cifar10.py
@@ -1,8 +1,8 @@
 import sklearn.cluster
 import numpy as np
 import autogp
-from autogp import likelihoods
-from autogp import kernels
+from autogp import lik
+from autogp import cov
 import tensorflow as tf
 from autogp import datasets
 from autogp import losses
@@ -64,8 +64,8 @@ if __name__ == '__main__':
     test = datasets.DataSet(test_X, test_Y)
 
     # Setup initial values for the model.
-    likelihood = likelihoods.Softmax()
-    kern = [kernels.RadialBasis(data.X.shape[1], lengthscale=10.0, input_scaling = IS_ARD) for i in range(10)]
+    likelihood = lik.Softmax()
+    kern = [cov.SquaredExponential(data.X.shape[1], length_scale=10.0, input_scaling = IS_ARD) for i in range(10)]
     # kern = [kernels.ArcCosine(X.shape[1], 2, 3, 5.0, 1.0, input_scaling=True) for i in range(10)] #RadialBasis(X.shape[1], input_scaling=True) for i in range(10)]
 
     Z = init_z(data.X, NUM_INDUCING)
