+++ experiments/mnist.py
@@ -1,7 +1,7 @@
 import sklearn.cluster
 import autogp
-from autogp import likelihoods
-from autogp import kernels
+from autogp import lik
+from autogp import cov
 import tensorflow as tf
 from autogp import datasets
 from autogp import losses
@@ -29,9 +29,9 @@ if __name__ == '__main__':
     data, test, _ = datasets.import_mnist()
 
     # Setup initial values for the model.
-    likelihood = likelihoods.Softmax()
-    kern = kernels.RadialBasis(data.X.shape[1], lengthscale=[10] * 10, std_dev=[1] * 10, white=[.01] * 10,
-                               input_scaling=IS_ARD)
+    likelihood = lik.Softmax()
+    kern = cov.SquaredExponential(data.X.shape[1], length_scale=[10] * 10, std_dev=[1] * 10, white=[.01] * 10,
+                                  input_scaling=IS_ARD)
     # kern = [kernels.ArcCosine(X.shape[1], 2, 3, 5.0, 1.0, input_scaling=True) for i in range(10)]
     # RadialBasis(X.shape[1], input_scaling=True) for i in xrange(10)]
